Django Simple Portfolio Management/requirements.txt
asgiref==3.7.2
asttokens==2.4.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
comm==0.2.0
crispy-bootstrap5==0.7
cryptography==41.0.4
customtkinter==5.2.1
darkdetect==0.8.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.6
django-allauth==0.57.0
django-crispy-forms==2.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
EasyProcess==1.1
entrypoint2==1.1
executing==2.0.1
greenlet==3.0.1
idna==3.4
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
Js2Py==0.74
jupyter_client==8.6.0
jupyter_core==5.5.0
keyboard==0.13.5
matplotlib-inline==0.1.6
MouseInfo==0.1.3
mss==9.0.1
nest-asyncio==1.5.8
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
pyjsparser==2.7.1
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python3-openid==3.2.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
wcwidth==0.2.12
